Deployable web service
