+++ docker-compose.yml
@@ -1,0 +1,10 @@
+version: '3'
+
+services:
+
+  wenling:
+    build:
+      context: ${PWD}
+      dockerfile: ./Dockerfile
+    ports:
+      - "8000:8000"
