+++ mypy.ini
@@ -1,4 +1,4 @@
 [mypy]
 files = .
-exclude = (wenling/tests/*|examples/*)
+exclude = (wenling/tests/*|wenling/scripts/*)
 ignore_missing_imports = True
