+++ pyproject.toml
@@ -21,6 +21,8 @@ bs4 = "^0.0.1"
 aiohttp = "^3.9.1"
 types-requests = "^2.31.0.10"
 uvicorn = "^0.24.0.post1"
+requests-oauthlib = "^1.3.1"
+flickrapi = "^2.4.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.3"
