+++ wenling/common/utils.py
@@ -138,80 +138,3 @@ class Logger:
         caller_name = caller_frame[3]
         caller_line = caller_frame[2]
         self.logger.warning(Fore.YELLOW + f"({caller_name} L{caller_line}): {message}" + Fore.RESET)
-
-
-@retrying.retry(wait_fixed=1000, stop_max_attempt_number=3)
-async def upload_image_to_imgur(image_path: str, logger: Logger, verbose: bool = False) -> str:
-    """
-    Uploads an image to Imgur and returns the URL of the uploaded image.
-
-    Parameters:
-        image_path (str): The path to the image file to be uploaded.
-        logger (Logger): The logger object for logging messages.
-        verbose (bool, optional): Whether to log verbose messages. Defaults to False.
-
-    Returns:
-        str: The URL of the uploaded image.
-
-    Raises:
-        ValueError: If IMGUR_CLIENT_ID environment variable is not set.
-    """
-    client_id = os.getenv("IMGUR_CLIENT_ID")
-    if not client_id:
-        raise ValueError("IMGUR_CLIENT_ID is not set")
-    url = "https://api.imgur.com/3/image"
-    headers = {"Authorization": f"Client-ID {client_id}"}
-
-    with open(image_path, "rb") as image_file:
-        image_data = image_file.read()
-
-    ssl_context = ssl.create_default_context()
-    ssl_context.check_hostname = False
-    ssl_context.verify_mode = ssl.CERT_NONE
-
-    async with aiohttp.ClientSession() as session:
-        try:
-            async with session.post(url, headers=headers, data={"image": image_data}, ssl=ssl_context) as response:
-                response.raise_for_status()
-                data = await response.json()
-        except Exception as e:
-            retest_interval = 300
-            if verbose:
-                logger.warning(f"Failed to upload image to imgur: {e}, retest in {retest_interval} seconds.")
-            await asyncio.sleep(retest_interval)
-            return ""
-
-    return data["data"]["link"]
-
-
-@retrying.retry(wait_fixed=1000, stop_max_attempt_number=3)
-async def save_image_to_imgur(image_url: str, logger: Logger, verbose: bool = False):
-    """
-    Save an image from the given URL to Imgur.
-
-    Args:
-        image_url (str): The URL of the image to be saved.
-        logger (Logger): An instance of the logger to log messages.
-        verbose (bool, optional): Whether to print verbose messages. Defaults to False.
-
-    Returns:
-        str: The URL of the image saved on Imgur.
-
-    Raises:
-        HTTPError: If there is an error while getting the image data.
-        Exception: If there is an error while uploading the image to Imgur.
-    """
-    # Get the image data
-    if image_url.startswith("https://github.com/"):
-        image_url = image_url.replace("blob", "raw")
-    response = requests.get(image_url)
-    response.raise_for_status()
-    # Create a temporary file and save the image data
-    with tempfile.NamedTemporaryFile(delete=False) as temp:
-        temp.write(response.content)
-        temp_file_path = temp.name
-    if verbose:
-        logger.info(f"Download file to {temp_file_path}.")
-    imgur_url = await upload_image_to_imgur(temp_file_path, logger, verbose=verbose)
-    os.remove(temp_file_path)
-    return imgur_url
